test_post() - fixed assertions that were incorrect.
test_post() - verified that the introductory email is sent to reviewers.

DSR 6762


test_review_mail_list() - filled in the test and removed the individual tests for add_to_list() and remove_from_list()


test_post_results() - verified that the rejection is not recorded if the reviewer opts to submit comments only.

DSR 6783


test_update_review_assignments() - to verify changes made for the DSR.

DSR 6786