Modified design_review_modification()
Added the cc_list parameter to pass in the old designer and/or peer.

Modified the code to add the designer, peer, and pcb input gate to the recipient list.

Modified the code the eliminate duplication in the cc and recipients list.  Common names were removed from the cc list.


DSR 7150