DSR #8122 - approval buttons disappeared.
The view.rhtml module was rendering the buttons, by invoking _review_selections,
only if comment_only wasn't set.
[ code = if !comment_only]
CCommented out the if, else and endif

The _review_selections.rhtml module was rendering the buttons only if there
was no response. Apparently "comment_only" is a response.
Commented out the if and endif.