DSR #9157 - handle large comments properly in Design Review updates
The program failed because flash was used to carry parameters through a
confirmation form when Reject was selected.

The confirmation form has been replaced with Javascripting on the page, triggered when
clicking on the Reject button.

This allowed simplification of the code. The "post_review" controller
method was moved into the reviewer_results method. The confirmation method and
its view were removed. Corresponding routes were also removed.

Application.js was modified to include jquery.dialog, which is used to
display the confirmer.