DSR #9354 Revisit prior change. The real issue was comparing report card count to outsource instruction count.  An assignment was cancelled in this design and the code wasn't considering that.
The code was using "assignment.completed?" with matched both
"Completed -> 1" and "Cancelled -> 2"

I added functions to count "Completed" and "Cancelled", then used
these appropriately for the functions use in testing if the final
review is ok to post.

I restored the expression used for the test since it has been working for
years.